Update dining_philosopher.c
The Dining Philosopher Problem:
The dining philosopher problem is one of the classical problems in the synchronization problem. It is a simple need to allocate several resources for several processes and it should not be in a deadlock and starvation manner. We can consider the dining philosopher problem as the simple representation of several processes and each process is requesting resources we need to allocate the resources for all the different processes such that deadlock never occurs and none of the processes stops anytime. Let us see a diagram to visualize the dining philosopher problem.